refactor: move folders and entities responsabilities
