+++ src/infra/config/database.go
@@ -5,7 +5,7 @@ import (
 	"log"
 	"os"
 
-	"github.com/rodrigoRVSN/beeus-api/src/domain/entities"
+	"github.com/rodrigoRVSN/beeus-api/src/models/users/entities"
 	"gorm.io/driver/postgres"
 	"gorm.io/gorm"
 	"gorm.io/gorm/logger"
