+++ src/main.go
@@ -2,7 +2,8 @@ package main
 
 import (
 	"github.com/gofiber/fiber/v2"
-	database "github.com/rodrigoRVSN/beeus-api/src/infra"
+	database "github.com/rodrigoRVSN/beeus-api/src/infra/config"
+	routes "github.com/rodrigoRVSN/beeus-api/src/infra/router"
 )
 
 func main() {
@@ -10,9 +11,7 @@ func main() {
 
 	app := fiber.New()
 
-	app.Get("/", func(context *fiber.Ctx) error {
-		return context.SendString("vasdjiasdhasioasdioasdjioasdjioasdjiopasj")
-	})
+	routes.SetupRoutes(app)
 
 	app.Listen(":4444")
 }
